# infra/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "GODS Platform Core",
  "description": "Sovereign AI governance backend for the G.O.D.S ecosystem.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/registry/models": {
   "get": {
    "tags": [
     "UDOC registry"
    ],
    "summary": "List Models",
    "operationId": "list_models_registry_models_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "UDOC registry"
    ],
    "summary": "Register Model",
    "operationId": "register_model_registry_models_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ModelReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/registry/models/{model_id}/status": {
   "post": {
    "tags": [
     "UDOC registry"
    ],
    "summary": "Set Status",
    "operationId": "set_status_registry_models__model_id__status_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "model_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Model Id"
      }
     },
     {
      "name": "new_status",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "New Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/seths/enrol": {
   "post": {
    "tags": [
     "SETHS"
    ],
    "summary": "Enrol",
    "operationId": "enrol_seths_enrol_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EnrolReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/seths/{ref}/advance": {
   "post": {
    "tags": [
     "SETHS"
    ],
    "summary": "Advance",
    "operationId": "advance_seths__ref__advance_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Ref"
      }
     },
     {
      "name": "monthly_value",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "default": 12000.0,
       "title": "Monthly Value"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/portal/student/register": {
   "post": {
    "tags": [
     "Portal \u00b7 Student"
    ],
    "summary": "Register",
    "operationId": "register_portal_student_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/StudentReg"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/portal/student/{student_ref}": {
   "get": {
    "tags": [
     "Portal \u00b7 Student"
    ],
    "summary": "Profile",
    "operationId": "profile_portal_student__student_ref__get",
    "parameters": [
     {
      "name": "student_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Student Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/portal/student/{student_ref}/progress": {
   "post": {
    "tags": [
     "Portal \u00b7 Student"
    ],
    "summary": "Update Progress",
    "operationId": "update_progress_portal_student__student_ref__progress_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "student_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Student Ref"
      }
     },
     {
      "name": "pct",
      "in": "query",
      "required": true,
      "schema": {
       "type": "number",
       "title": "Pct"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/portal/student/{student_ref}/apply/{opportunity_id}": {
   "post": {
    "tags": [
     "Portal \u00b7 Student"
    ],
    "summary": "Apply",
    "operationId": "apply_portal_student__student_ref__apply__opportunity_id__post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "student_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Student Ref"
      }
     },
     {
      "name": "opportunity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Opportunity Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/portal/employee/{employee_ref}": {
   "get": {
    "tags": [
     "Portal \u00b7 Employee"
    ],
    "summary": "Profile",
    "operationId": "profile_portal_employee__employee_ref__get",
    "parameters": [
     {
      "name": "employee_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Employee Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/portal/employee/{employee_ref}/timesheet": {
   "post": {
    "tags": [
     "Portal \u00b7 Employee"
    ],
    "summary": "Submit Timesheet",
    "operationId": "submit_timesheet_portal_employee__employee_ref__timesheet_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "employee_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Employee Ref"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TimesheetReq"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/portal/employee/timesheet/{timesheet_id}/approve": {
   "post": {
    "tags": [
     "Portal \u00b7 Employee"
    ],
    "summary": "Approve",
    "operationId": "approve_portal_employee_timesheet__timesheet_id__approve_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "timesheet_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Timesheet Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/portal/employee/{employee_ref}/payslip": {
   "get": {
    "tags": [
     "Portal \u00b7 Employee"
    ],
    "summary": "Payslip",
    "operationId": "payslip_portal_employee__employee_ref__payslip_get",
    "parameters": [
     {
      "name": "employee_ref",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Employee Ref"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "EnrolReq": {
    "properties": {
     "qualification": {
      "type": "string",
      "title": "Qualification",
      "default": "Software Developer 118707"
     },
     "nqf_level": {
      "type": "integer",
      "title": "Nqf Level",
      "default": 5
     },
     "count": {
      "type": "integer",
      "title": "Count",
      "default": 1
     }
    },
    "type": "object",
    "title": "EnrolReq"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "ModelReq": {
    "properties": {
     "model_id": {
      "type": "string",
      "title": "Model Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "operator_id": {
      "type": "string",
      "title": "Operator Id"
     },
     "risk_tier": {
      "type": "string",
      "title": "Risk Tier",
      "default": "NOTABLE"
     },
     "use_case": {
      "type": "string",
      "title": "Use Case",
      "default": ""
     },
     "jurisdiction": {
      "type": "string",
      "title": "Jurisdiction",
      "default": "ZA"
     }
    },
    "type": "object",
    "required": [
     "model_id",
     "name",
     "operator_id"
    ],
    "title": "ModelReq"
   },
...
   "StudentReg": {
    "properties": {
     "full_name": {
      "type": "string",
      "title": "Full Name"
     },
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "qualification": {
      "type": "string",
      "title": "Qualification",
      "default": "Software Developer 118707"
     },
     "nqf_level": {
      "type": "integer",
      "title": "Nqf Level",
      "default": 5
     }
    },
    "type": "object",
    "required": [
     "full_name",
     "email"
    ],
    "title": "StudentReg"
   },
   "TimesheetReq": {
    "properties": {
     "hours": {
      "type": "number",
      "title": "Hours"
     },
     "period": {
      "type": "string",
      "title": "Period",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "hours"
    ],
    "title": "TimesheetReq"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "auth/login"
     }
    }
   }
  }
 }
}